Removed the scary placeholders + refactor captions package
