+++ app/src/main/java/com/storyspots/cache/StoryListCache.kt
@@ -3,15 +3,15 @@ package com.storyspots.cache
 import android.content.Context
 import android.util.Log
 import android.util.LruCache
-import com.storyspots.caption.StoryData
-import com.storyspots.caption.CachedStoryData
+import com.storyspots.caption.model.StoryData
+import com.storyspots.caption.model.CachedStoryData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
 import java.io.ObjectInputStream
 import java.io.ObjectOutputStream
 
-class StoryCache(private val context: Context) {
+class StoryListCache(private val context: Context) {
     private val cacheFile = File(context.cacheDir, "stories_cache.dat")
     private val memoryCacheSize = 50
     private val memoryCache = LruCache<String, StoryData>(memoryCacheSize)
@@ -51,16 +51,6 @@ class StoryCache(private val context: Context) {
             Log.d("StoryCache", "Successfully cached ${stories.size} stories")
         } catch (e: Exception) {
             Log.e("StoryCache", "Failed to write cache", e)
-        }
-    }
-
-    fun clearCache() {
-        try {
-            cacheFile.delete()
-            memoryCache.evictAll()
-            Log.d("StoryCache", "Cache cleared")
-        } catch (e: Exception) {
-            Log.e("StoryCache", "Failed to clear cache", e)
         }
     }
 }
