+++ app/src/main/java/com/storyspots/caption/model/UserModel.kt
@@ -1,0 +1,7 @@
+package com.storyspots.caption.model
+
+data class UserData(
+    val id: String = "",
+    val username: String = "Deleted User",
+    val profileImageUrl: String = ""
+)
