+++ app/src/main/java/com/storyspots/ui/theme/Color.kt
@@ -10,7 +10,7 @@ val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
 
-val Pink = Color(0xFFFF2D87)
+val Pink = Color(0xFFFF398F)
 val LightPink = Color(0xFFFFD6E5)
 val UltraLightPink =  Color(0xFFFFE5EC)
 val DarkText = Color(0xFF1D1D1D)
@@ -25,3 +25,4 @@ val HintGray = Color(0xFF575764)
 val LightGray = Color(0xA4C8CCCC)
 val ErrorColour = Color(0xFFa50001)
 val SuccessColour = Color(0xFF4CBB17)
+val NeonGreen = Color(0xFF8FFF53)
