+++ app/src/main/java/com/storyspots/yourFeed/YourFeedScreen.kt
@@ -1,21 +1,16 @@
 package com.storyspots.yourFeed
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
@@ -23,11 +18,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
 import com.storyspots.R
-import com.storyspots.caption.StoryData
-import com.storyspots.yourFeed.YourFeedViewModel.Companion.formatFirebaseTimestamp
+import com.storyspots.caption.model.StoryData
 import com.storyspots.yourFeed.components.DeleteConfirmationDialog
 import com.storyspots.yourFeed.components.StoryCard
 
@@ -43,7 +35,6 @@ fun YourFeedScreen(
     var showDeleteDialog by remember { mutableStateOf(false) }
     var storyToDelete by remember { mutableStateOf<StoryData?>(null) }
 
-    // Handle delete state changes
     LaunchedEffect(deleteState) {
         when (deleteState) {
             is DeleteState.Success -> {
@@ -55,7 +46,6 @@ fun YourFeedScreen(
         }
     }
 
-    // Delete confirmation dialog
     if (showDeleteDialog && storyToDelete != null) {
         DeleteConfirmationDialog(
             onDismiss = {
@@ -87,7 +77,6 @@ fun YourFeedScreen(
                 )
             },
             actions = {
-                // Refresh button for error states
                 if (uiState is YourFeedUiState.Error) {
                     IconButton(onClick = { viewModel.retry() }) {
                         Icon(
@@ -103,7 +92,6 @@ fun YourFeedScreen(
             )
         )
 
-        // Show delete error as snackbar
         if (deleteState is DeleteState.Error) {
             Card(
                 modifier = Modifier
@@ -121,7 +109,6 @@ fun YourFeedScreen(
             }
         }
 
-        // Main content based on UI state
         when (uiState) {
             is YourFeedUiState.Loading -> {
                 LoadingContent()
